requirements.txt
amqp==5.0.1
appdirs==1.4.4
asgiref==3.2.10
beautifulsoup4==4.9.3
billiard==3.6.3.0
bs4==0.0.1
celery==5.0.0
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
click-didyoumean==0.0.3
click-repl==0.1.6
cssselect==1.1.0
dj-database-url==0.5.0
Django==3.1.2
django-heroku==0.3.1
fake-useragent==0.1.11
gunicorn==20.0.4
idna==2.10
importlib-metadata==2.0.0
kombu==5.0.2
lxml==4.6.0
parse==1.18.0
prompt-toolkit==3.0.8
psycopg2==2.7.5
pyee==7.0.4
pyppeteer==0.2.2
pyquery==1.4.1
python-dateutil==2.8.1
pytz==2020.1
redis==3.5.3
requests==2.24.0
requests-html==0.10.0
six==1.15.0
soupsieve==2.0.1
sqlparse==0.4.1
tqdm==4.50.2
urllib3==1.25.10
vine==5.0.0
w3lib==1.22.0
wcwidth==0.2.5
websockets==8.1
whitenoise==5.2.0
zipp==3.3.1